feat(rendezvous-server): Build in CI; fix timeout issue
