+++ libp2p-rendezvous-server/src/main.rs
@@ -1,34 +1,26 @@
 use anyhow::{Context, Result};
-use futures::{AsyncRead, AsyncWrite, StreamExt};
-use libp2p::core::muxing::StreamMuxerBox;
-use libp2p::core::transport::Boxed;
-use libp2p::core::upgrade::Version;
-use libp2p::identity::ed25519;
-use libp2p::noise;
-use libp2p::rendezvous::server::Behaviour;
+use futures::StreamExt;
+use libp2p::identity::{self, ed25519};
+use libp2p::rendezvous;
 use libp2p::swarm::SwarmEvent;
-use libp2p::tcp;
-use libp2p::yamux;
-use libp2p::{dns, SwarmBuilder};
-use libp2p::{identity, rendezvous, Multiaddr, PeerId, Swarm, Transport};
-use libp2p_tor::{AddressConversion, TorTransport};
-use std::fmt;
 use std::path::{Path, PathBuf};
-use std::time::Duration;
 use structopt::StructOpt;
 use tokio::fs;
 use tokio::fs::{DirBuilder, OpenOptions};
 use tokio::io::AsyncWriteExt;
-use tor_hsservice::config::OnionServiceConfigBuilder;
 use tracing::level_filters::LevelFilter;
 use tracing_subscriber::FmtSubscriber;
+
+use crate::swarm::{create_swarm, create_swarm_with_onion, Addresses};
+
+pub mod swarm;
 
 #[derive(Debug, StructOpt)]
 struct Cli {
     /// Path to the file that contains the secret key of the rendezvous server's
     /// identity keypair
     /// If the file does not exist, a new secret key will be generated and saved to the file
-    #[structopt(long, default_value = "rendezvous-server-secret.key")]
+    #[structopt(long, default_value = "./rendezvous-server-secret.key")]
     secret_file: PathBuf,
 
     /// Port used for listening on TCP (default)
@@ -37,7 +29,7 @@ struct Cli {
 
     /// Enable listening on Tor onion service
     #[structopt(long)]
-    onion: bool,
+    no_onion: bool,
 
     /// Port for the onion service (only used if --onion is enabled)
     #[structopt(long, default_value = "8888")]
@@ -57,16 +49,16 @@ struct Cli {
 async fn main() -> Result<()> {
     let cli = Cli::from_args();
 
-    init_tracing(LevelFilter::INFO, cli.json, cli.no_timestamp);
+    init_tracing(LevelFilter::TRACE, cli.json, cli.no_timestamp);
 
     let secret_key = load_secret_key_from_file(&cli.secret_file).await?;
 
     let identity = identity::Keypair::from(ed25519::Keypair::from(secret_key));
 
-    let mut swarm = if cli.onion {
-        create_swarm_with_onion(identity, cli.onion_port).await?
-    } else {
+    let mut swarm = if cli.no_onion {
         create_swarm(identity)?
+    } else {
+        create_swarm_with_onion(identity, cli.onion_port).await?
     };
 
     tracing::info!(peer_id=%swarm.local_peer_id(), "Rendezvous server peer id");
@@ -126,7 +118,42 @@ fn init_tracing(level: LevelFilter, json_format: bool, no_timestamp: bool) {
     let is_terminal = atty::is(atty::Stream::Stderr);
 
     let builder = FmtSubscriber::builder()
-        .with_env_filter(format!("rendezvous_server={}", level))
+        .with_env_filter(format!(
+            "rendezvous_server={},\
+                 libp2p={},\
+                 libp2p_allow_block_list={},\
+                 libp2p_connection_limits={},\
+                 libp2p_core={},\
+                 libp2p_dns={},\
+                 libp2p_identity={},\
+                 libp2p_noise={},\
+                 libp2p_ping={},\
+                 libp2p_rendezvous={},\
+                 libp2p_request_response={},\
+                 libp2p_swarm={},\
+                 libp2p_tcp={},\
+                 libp2p_tls={},\
+                 libp2p_tor={},\
+                 libp2p_websocket={},\
+                 libp2p_yamux={}",
+            level,
+            level,
+            level,
+            level,
+            level,
+            level,
+            level,
+            level,
+            level,
+            level,
+            level,
+            level,
+            level,
+            level,
+            level,
+            level,
+            level
+        ))
         .with_writer(std::io::stderr)
         .with_ansi(is_terminal)
         .with_target(false);
@@ -192,128 +219,4 @@ async fn write_secret_key_to_file(secret_key: &ed25519::SecretKey, path: PathBuf
     file.write_all(secret_key.as_ref()).await?;
 
     Ok(())
-}
-
-fn create_swarm(identity: identity::Keypair) -> Result<Swarm<Behaviour>> {
-    let transport = create_transport(&identity).context("Failed to create transport")?;
-    let rendezvous = rendezvous::server::Behaviour::new(rendezvous::server::Config::default());
-
-    let swarm = SwarmBuilder::with_existing_identity(identity)
-        .with_tokio()
-        .with_other_transport(|_| transport)?
-        .with_behaviour(|_| rendezvous)?
-        .build();
-
-    Ok(swarm)
-}
-
-async fn create_swarm_with_onion(
-    identity: identity::Keypair,
-    onion_port: u16,
-) -> Result<Swarm<Behaviour>> {
-    let (transport, onion_address) = create_transport_with_onion(&identity, onion_port)
-        .await
-        .context("Failed to create transport with onion")?;
-    let rendezvous = rendezvous::server::Behaviour::new(rendezvous::server::Config::default());
-
-    let mut swarm = SwarmBuilder::with_existing_identity(identity)
-        .with_tokio()
-        .with_other_transport(|_| transport)?
-        .with_behaviour(|_| rendezvous)?
-        .build();
-
-    // Listen on the onion address
-    swarm
-        .listen_on(onion_address.clone())
-        .context("Failed to listen on onion address")?;
-
-    tracing::info!(%onion_address, "Onion service configured");
-
-    Ok(swarm)
-}
-
-fn create_transport(identity: &identity::Keypair) -> Result<Boxed<(PeerId, StreamMuxerBox)>> {
-    let tcp = tcp::tokio::Transport::new(tcp::Config::new().nodelay(true));
-    let tcp_with_dns = dns::tokio::Transport::system(tcp)?;
-
-    let transport = authenticate_and_multiplex(tcp_with_dns.boxed(), &identity).unwrap();
-
-    Ok(transport)
-}
-
-async fn create_transport_with_onion(
-    identity: &identity::Keypair,
-    onion_port: u16,
-) -> Result<(Boxed<(PeerId, StreamMuxerBox)>, Multiaddr)> {
-    // Create TCP transport
-    let tcp = tcp::tokio::Transport::new(tcp::Config::new().nodelay(true));
-    let tcp_with_dns = dns::tokio::Transport::system(tcp)?;
-
-    // Create Tor transport
-    let mut tor_transport = TorTransport::unbootstrapped()
-        .await?
-        .with_address_conversion(AddressConversion::IpAndDns);
-
-    // Create onion service configuration
-    let onion_service_config = OnionServiceConfigBuilder::default()
-        .nickname(
-            identity
-                .public()
-                .to_peer_id()
-                .to_base58()
-                .to_ascii_lowercase()
-                .parse()
-                .unwrap(),
-        )
-        .num_intro_points(3)
-        .build()
-        .unwrap();
-
-    // Add onion service and get the address
-    let onion_address = tor_transport.add_onion_service(onion_service_config, onion_port)?;
-
-    // Combine transports
-    let combined_transport = tcp_with_dns
-        .boxed()
-        .or_transport(tor_transport.boxed())
-        .boxed();
-
-    let transport = authenticate_and_multiplex(combined_transport, &identity).unwrap();
-
-    Ok((transport, onion_address))
-}
-
-fn authenticate_and_multiplex<T>(
-    transport: Boxed<T>,
-    identity: &identity::Keypair,
-) -> Result<Boxed<(PeerId, StreamMuxerBox)>>
-where
-    T: AsyncRead + AsyncWrite + Unpin + Send + 'static,
-{
-    let noise_config = noise::Config::new(identity).unwrap();
-
-    let transport = transport
-        .upgrade(Version::V1)
-        .authenticate(noise_config)
-        .multiplex(yamux::Config::default())
-        .timeout(Duration::from_secs(20))
-        .map(|(peer, muxer), _| (peer, StreamMuxerBox::new(muxer)))
-        .boxed();
-
-    Ok(transport)
-}
-
-struct Addresses<'a>(&'a [Multiaddr]);
-
-// Prints an array of multiaddresses as a comma seperated string
-impl fmt::Display for Addresses<'_> {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        let display = self
-            .0
-            .iter()
-            .map(|addr| addr.to_string())
-            .collect::<Vec<String>>()
-            .join(",");
-        write!(f, "{}", display)
-    }
 }
